+++ lib/haml/attribute_builder.rb
@@ -15,11 +15,7 @@ module Haml
             data_attributes = attributes.delete(key)
             data_attributes = flatten_data_attributes(data_attributes, '', join_char)
             data_attributes = build_data_keys(data_attributes, hyphenate_data_attrs, key)
-            data_attributes.keys.each do |key|
-              if key =~ INVALID_ATTRIBUTE_NAME_REGEX
-                raise InvalidAttributeNameError.new("Invalid attribute name '#{key}' was built!")
-              end
-            end
+            verify_attribute_names!(data_attributes.keys)
             attributes = data_attributes.merge(attributes)
           end
         end
@@ -95,6 +91,14 @@ module Haml
       def merge_values(key, *values)
         values.inject(nil) do |to, from|
           merge_value(key, to, from)
+        end
+      end
+
+      def verify_attribute_names!(attribute_names)
+        attribute_names.each do |attribute_name|
+          if attribute_name =~ INVALID_ATTRIBUTE_NAME_REGEX
+            raise InvalidAttributeNameError.new("Invalid attribute name '#{attribute_name}' was rendered")
+          end
         end
       end